requirements.txt
python-telegram-bot==20.*
python-telegram-bot[socks]
requests
python-dotenv